web/requirements.txt
django-crispy-forms==1.8.0
django-storages[google]==1.7.2
gevent==1.4.0
google-auth==1.6.3
google-cloud-container==0.3.0
gunicorn==19.9.0
kubernetes==10.0.1
psycopg2==2.7.7
pyyaml==5.1.2
sentry-sdk==0.12.3
